# Conflicts:
#	src/main/kotlin/com/fastcampus/commerce/admin/order/application/AdminOrderService.kt
#	src/main/kotlin/com/fastcampus/commerce/admin/order/infrastructure/controller/AdminOrderController.kt
#	src/main/kotlin/com/fastcampus/commerce/admin/order/infrastructure/request/AdminOrderSearchRequest.kt
#	src/main/kotlin/com/fastcampus/commerce/admin/order/infrastructure/response/AdminOrderListResponse.kt
#	src/main/kotlin/com/fastcampus/commerce/admin/order/interfaces/AdminOrderQuery.kt
#	src/main/kotlin/com/fastcampus/commerce/admin/order/interfaces/AdminOrderQueryImpl.kt

+++ src/main/kotlin/com/fastcampus/commerce/review/application/response/ProductReviewRatingResponse.kt
@@ -13,11 +13,11 @@ data class ProductReviewRatingResponse(
                 averageRating = response.average,
                 totalCount = response.totalCount,
                 ratingDistribution = RatingDistributionResponse(
-                    oneStarCount = response.oneStarCount,
-                    twoStarsCount = response.twoStarsCount,
-                    threeStarsCount = response.threeStarsCount,
-                    fourStarsCount = response.fourStarsCount,
-                    fiveStarsCount = response.fiveStarsCount,
+                    oneStarCount = response.oneStarCount.toInt(),
+                    twoStarsCount = response.twoStarsCount.toInt(),
+                    threeStarsCount = response.threeStarsCount.toInt(),
+                    fourStarsCount = response.fourStarsCount.toInt(),
+                    fiveStarsCount = response.fiveStarsCount.toInt(),
                 ),
             )
     }
